fix(mbtiles): add web stub via conditional imports, fix unused imports
- Add `MbTilesTileProvider` web stub (`mbtiles_tile_provider_web.dart`)
  that throws `UnsupportedError` on web; avoids importing `mbtiles`/
  `sqlite3` which cannot compile for web with WASM bindings
- Update `flutter_map_mbtiles.dart` to use conditional export:
  `if (dart.library.html) 'src/mbtiles_tile_provider_web.dart'`
- Split example `flutter_map_mbtiles/page.dart` and
  `vector_map_tiles_mbtiles/page.dart` into `_io`/`_web` pairs;
  web builds show "MBTiles is not supported on web."
- Remove unnecessary `latlong2` imports from example pages and test
  utility (`mbtiles` 0.5.0 re-exports `LatLng` via its own barrel)
