+++ flutter_map_mbtiles/lib/flutter_map_mbtiles.dart
@@ -1,1 +1,2 @@
-export 'src/mbtiles_tile_provider.dart';
+export 'src/mbtiles_tile_provider.dart'
+    if (dart.library.html) 'src/mbtiles_tile_provider_web.dart';
