+++ flutter_map_mbtiles/test/utils/test_app.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map_mbtiles/src/mbtiles_tile_provider.dart';
-import 'package:latlong2/latlong.dart';
 import 'package:mbtiles/mbtiles.dart';
 
 class TestApp extends StatelessWidget {
